topology: extract PSD/MGD/PGD + solver + mapper-utils + FSD->PSD into TT::ScaleoutTopology
Create a lean, runtime-free static library TT::ScaleoutTopology holding the topology
solver, mapper UTILS, PSD/MGD/PGD (proto + class), and the offline FSD->PSD builder.
mesh_graph and the runtime TopologyMapper class stay in fabric; fabric links the new lib.

- fsd_to_psd.hpp promoted to tt-metalium API tree (<tt-metalium/experimental/fabric/
  fsd_to_psd.hpp>) + proto-free build_psd_from_fsd_file(path) entry point.
- The 3 MeshGraph-coupled functions: build_adjacency_map_logical(MeshGraph) had no
  callers -> deleted; build_adjacency_graph_logical(MeshGraph) and
  build_logical_multi_mesh_adjacency_graph(MeshGraph) relocated into topology_mapper.cpp
  (no new file), forwarding to an exposed "parts" overload in topology_mapper_utils.
- Dropped dead heavy includes (metal_context/tt_cluster) from the moved TUs; inlined a
  local get_host_name in physical_system_descriptor.cpp.
- CMake: split PSD/MGD/PGD protos into the topology lib (hidden visibility); fabric links
  TT::ScaleoutTopology; generate_rank_bindings + test_fsd_to_psd link it explicitly.

Verified: cmake --build scaleout_topology test_fsd_to_psd fabric generate_rank_bindings
all link clean; libtt_metal.so rebuilds; test_fsd_to_psd links with NO libtt_metal and
5/5 gtests pass.
