+++ tools/scaleout/CMakeLists.txt
@@ -132,49 +132,9 @@ tt_reuse_precompile_headers(scaleout_tools TT::CommonPCH)
 
 install(TARGETS scaleout_tools EXPORT Metalium LIBRARY COMPONENT metalium-runtime FILE_SET api COMPONENT metalium-dev)
 
-# ---------------------------------------------------------------------------
-# fsd_to_psd: lean, offline FactorySystemDescriptor -> PhysicalSystemDescriptor conversion.
-#
-# Deliberately does NOT link the tt-metal runtime (Metalium). It links only TT::ScaleoutTools, which supplies
-# the FSD proto and the board library and is itself runtime-free. The PSD proto (tt::fabric::proto, compiled
-# once inside the `fabric` target / libtt_metal) is consumed here by HEADER only: the FSD->PSD builder emits
-# PSD-proto messages, and the PSD-proto *symbols* are resolved by whatever binary links this lib alongside
-# tt_metal (e.g. generate_rank_bindings). This keeps the PSD proto compiled exactly once — avoiding protobuf
-# "File already exists in database" duplicate-descriptor aborts — while keeping this library metal-runtime-free.
-add_library(fsd_to_psd STATIC)
-add_library(TT::FsdToPsd ALIAS fsd_to_psd)
-target_sources(fsd_to_psd PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/fsd_to_psd/fsd_to_psd.cpp)
-
-# Public: consumers include "fsd_to_psd/fsd_to_psd.hpp". That header pulls in the FSD and PSD generated proto
-# headers, so a consumer must also have those generated-header dirs on its include path (generate_rank_bindings
-# already adds ${Metalium_BINARY_DIR}/tools/scaleout and fabric's BINARY_DIR).
-target_include_directories(fsd_to_psd PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
-target_include_directories(
-    fsd_to_psd
-    SYSTEM
-    PRIVATE
-        ${CMAKE_CURRENT_BINARY_DIR} # scaleout-generated protobuf/factory_system_descriptor.pb.h
-        "$<TARGET_PROPERTY:fabric,BINARY_DIR>" # fabric-generated protobuf/physical_system_descriptor.pb.h
-)
-
-# Headers only: the FSD proto is generated by scaleout_tools, the PSD proto by fabric. We depend on both custom
-# targets for build ordering but do NOT compile either .pb.cc here (symbols come from TT::ScaleoutTools and,
-# at final link, from tt_metal/fabric respectively).
-add_dependencies(fsd_to_psd scaleout_generated_protos fabric_generated_protos)
-
-target_link_libraries(
-    fsd_to_psd
-    PUBLIC
-        TT::ScaleoutTools
-    PRIVATE
-        protobuf::libprotobuf
-        fmt::fmt-header-only
-        tt-logger::tt-logger
-)
-target_compile_features(fsd_to_psd PUBLIC cxx_std_20)
-
-# No install(): fsd_to_psd is a build-time static lib; its objects are linked directly into consumers such as
-# generate_rank_bindings. Add an install/export here if it ever needs to ship as a standalone artifact.
+# NOTE: the offline FSD->PSD builder (tools/scaleout/fsd_to_psd/fsd_to_psd.cpp) is compiled into the lean
+# TT::ScaleoutTopology static library (tt_metal/fabric/CMakeLists.txt), alongside PSD/MGD/PGD, the topology
+# solver, and the topology mapper utils, so all the runtime-free topology code lives in one library.
 
 add_executable(2d_big_mesh_cabling_gen)
 target_sources(2d_big_mesh_cabling_gen PRIVATE ${SCALEOUT_2D_BIG_MESH_CABLING_GEN_SRCS})
@@ -386,6 +346,7 @@ if(ENABLE_DISTRIBUTED)
         generate_rank_bindings
         PRIVATE
             scaleout_tools
+            TT::ScaleoutTopology # PSD/MGD/PGD/solver/mapper-utils/FSD->PSD symbols (hidden in libtt_metal)
             tt_metal
             OpenMPI::MPI
             protobuf::libprotobuf
