+++ tt_metal/fabric/CMakeLists.txt
@@ -32,46 +32,105 @@ target_sources(
 
 # Include helper functions and generate headers from protobuf schemas
 include(protobuf)
-set(PROTO_SCHEMAS
+
+# Descriptor protos consumed by the runtime-free topology lib vs. fabric-runtime protos. All generate into the
+# same protobuf/ dir so consumers resolve headers via ${CMAKE_CURRENT_BINARY_DIR}.
+set(TOPOLOGY_PROTO_SCHEMAS
     ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/mesh_graph_descriptor.proto
     ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/physical_system_descriptor.proto
+    ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/physical_grouping_descriptor.proto
+)
+set(FABRIC_PROTO_SCHEMAS
     ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/router_port_directions.proto
     ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/port_descriptor_table.proto
     ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/intermesh_connection_table.proto
-    ${CMAKE_CURRENT_SOURCE_DIR}/protobuf/physical_grouping_descriptor.proto
 )
 
-# Collect all generated protobuf files
-set(GENERATED_PROTO_SRCS)
-foreach(PROTO_FILE ${PROTO_SCHEMAS})
+set(TOPOLOGY_GENERATED_PROTO_SRCS)
+foreach(PROTO_FILE ${TOPOLOGY_PROTO_SCHEMAS})
     GENERATE_PROTO_FILES(${PROTO_FILE} OUTPUT_DIR ${CMAKE_CURRENT_BINARY_DIR})
-    list(APPEND GENERATED_PROTO_SRCS ${PROTO_SRCS})
+    list(APPEND TOPOLOGY_GENERATED_PROTO_SRCS ${PROTO_SRCS})
 endforeach()
 
-# Expose a custom target for the generated protobuf files so that external consumers
-# can depend on them directly without going through the fabric OBJECT library.
-add_custom_target(fabric_generated_protos DEPENDS ${GENERATED_PROTO_SRCS})
-add_dependencies(fabric fabric_generated_protos)
+set(FABRIC_GENERATED_PROTO_SRCS)
+foreach(PROTO_FILE ${FABRIC_PROTO_SCHEMAS})
+    GENERATE_PROTO_FILES(${PROTO_FILE} OUTPUT_DIR ${CMAKE_CURRENT_BINARY_DIR})
+    list(APPEND FABRIC_GENERATED_PROTO_SRCS ${PROTO_SRCS})
+endforeach()
+
+# ---------------------------------------------------------------------------
+# TT::ScaleoutTopology — lean, runtime-free static library: MGD + PSD + PGD (proto + class), the topology solver,
+# the topology mapper UTILS, and the offline FSD->PSD builder. mesh_graph + the runtime TopologyMapper class stay
+# in fabric. Pulls tt-metalium headers via include-dirs only (no libtt_metal link), so FM/offline tools can link
+# just this lib without the metal runtime.
+# ---------------------------------------------------------------------------
+add_library(scaleout_topology STATIC ${TOPOLOGY_SOURCES} ${TOPOLOGY_GENERATED_PROTO_SRCS})
+add_library(TT::ScaleoutTopology ALIAS scaleout_topology)
+set_source_files_properties(${TOPOLOGY_GENERATED_PROTO_SRCS} PROPERTIES SKIP_UNITY_BUILD_INCLUSION TRUE)
+
+add_custom_target(topology_generated_protos DEPENDS ${TOPOLOGY_GENERATED_PROTO_SRCS})
+add_dependencies(scaleout_topology topology_generated_protos scaleout_generated_protos)
+if(TARGET all_generated_files)
+    add_dependencies(all_generated_files topology_generated_protos)
+endif()
+
+target_include_directories(
+    scaleout_topology
+    PRIVATE
+        "$<TARGET_PROPERTY:TT::Metalium,INCLUDE_DIRECTORIES>" # tt-metalium public headers (incl fsd_to_psd.hpp)
+        ${CMAKE_CURRENT_SOURCE_DIR}
+        ${CMAKE_CURRENT_SOURCE_DIR}/../../tools/scaleout # board/board.hpp
+    SYSTEM
+    PRIVATE
+        ${CMAKE_CURRENT_BINARY_DIR} # this lib's generated proto headers
+        ${Metalium_BINARY_DIR}/tools/scaleout # scaleout-generated FSD proto header (for fsd_to_psd)
+)
+target_link_libraries(
+    scaleout_topology
+    PUBLIC
+        TT::STL
+        umd::tt-umd
+    PRIVATE
+        TT::ScaleoutTools
+        TT::Metalium::HostDevCommon
+        enchantum::enchantum
+        fmt::fmt-header-only
+        tt-logger::tt-logger
+        yaml-cpp::yaml-cpp
+        protobuf::libprotobuf
+        cadical::cadical
+)
+target_compile_features(scaleout_topology PUBLIC cxx_std_20)
+# Hidden visibility (same as TT::ScaleoutTools): this lib carries protobuf message classes, so its descriptor
+# symbols must be hidden to avoid cross-module "File already exists in database" aborts. Consumers reaching these
+# symbols only transitively through libtt_metal.so must therefore link TT::ScaleoutTopology explicitly.
+set_target_properties(
+    scaleout_topology
+    PROPERTIES
+        CXX_VISIBILITY_PRESET
+            "hidden"
+        VISIBILITY_INLINES_HIDDEN
+            TRUE
+)
+
+# --- fabric: remaining runtime protos + link the topology lib -------
+add_custom_target(fabric_generated_protos DEPENDS ${FABRIC_GENERATED_PROTO_SRCS})
+add_dependencies(fabric fabric_generated_protos scaleout_topology)
 if(TARGET all_generated_files)
     add_dependencies(all_generated_files fabric_generated_protos)
 endif()
 
-# Exclude protobuf generated files from unity builds (they have conflicting static symbols)
-set_source_files_properties(
-    ${GENERATED_PROTO_SRCS}
-    PROPERTIES
-        SKIP_UNITY_BUILD_INCLUSION
-            TRUE
-)
+set_source_files_properties(${FABRIC_GENERATED_PROTO_SRCS} PROPERTIES SKIP_UNITY_BUILD_INCLUSION TRUE)
 
-# Add the generated protobuf files to your target
-target_sources(fabric PRIVATE ${GENERATED_PROTO_SRCS})
+target_sources(fabric PRIVATE ${FABRIC_GENERATED_PROTO_SRCS})
 
 target_include_directories(fabric PRIVATE .)
 target_include_directories(fabric SYSTEM PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
 
 target_link_libraries(
     fabric
+    PUBLIC
+        TT::ScaleoutTopology
     PRIVATE
         TT::ScaleoutTools
         Metalium::Metal::LLRT
